Importing logs moved from Logs analysis to Importers
