+++ importers/src/main/java/pl/edu/icm/coansys/importers/logs/AuditEntry2Protos.java
@@ -1,7 +1,7 @@
 /*
  * (C) 2010-2012 ICM UW. All rights reserved.
  */
-package pl.edu.icm.coansys.logsanalysis.transformers;
+package pl.edu.icm.coansys.importers.logs;
 
 import java.util.Date;
 import java.util.regex.Matcher;
