+++ importers/src/main/java/pl/edu/icm/coansys/importers/logs/AuditServiceToRemoteHTable.java
@@ -1,7 +1,7 @@
 /*
  * (C) 2010-2012 ICM UW. All rights reserved.
  */
-package pl.edu.icm.coansys.logsanalysis.logsacquisition;
+package pl.edu.icm.coansys.importers.logs;
 
 import java.io.IOException;
 import java.util.List;
@@ -12,7 +12,6 @@ import org.apache.hadoop.hbase.rest.client.RemoteHTable;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 import pl.edu.icm.coansys.importers.models.LogsProtos.LogsMessage;
-import pl.edu.icm.coansys.logsanalysis.transformers.AuditEntry2Protos;
 import pl.edu.icm.synat.api.services.SynatServiceRef;
 import pl.edu.icm.synat.api.services.audit.AuditService;
 import pl.edu.icm.synat.api.services.audit.model.AuditEntry;
