+++ importers/src/main/java/pl/edu/icm/coansys/importers/logs/GenerateDummyLogs.java
@@ -1,7 +1,7 @@
 /*
  * (C) 2010-2012 ICM UW. All rights reserved.
  */
-package pl.edu.icm.coansys.logsanalysis.logsacquisition;
+package pl.edu.icm.coansys.importers.logs;
 
 import java.io.IOException;
 import java.io.Serializable;
@@ -10,9 +10,6 @@ import java.net.MalformedURLException;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.*;
-import pl.edu.icm.coansys.logsanalysis.constants.ParamNames;
-import pl.edu.icm.coansys.logsanalysis.transformers.AuditEntry2Protos;
-import pl.edu.icm.coansys.logsanalysis.transformers.BytesArray2SequenceFile;
 import pl.edu.icm.synat.api.services.audit.model.AuditEntry;
 
 /**
@@ -101,12 +98,12 @@ public final class GenerateDummyLogs {
                 String eventType = EVENTTYPES[random.nextInt(EVENTTYPES.length)];
 
                 List<String> argsList = new ArrayList<String>();
-                argsList.add("[" + ParamNames.RESOURCE_ID_PARAM + "=" + RESOURCES[random.nextInt(RESOURCES.length)] + "]");
-                argsList.add("[" + ParamNames.SESSION_ID_PARAM + "=" + sessionId + "]");
-                argsList.add("[" + ParamNames.IP_PARAM + "=" + IPADDRESSES[random.nextInt(IPADDRESSES.length)] + "]");
-                argsList.add("[" + ParamNames.URL_PARAM + "=" + URLS[random.nextInt(URLS.length)] + "]");
-                argsList.add("[" + ParamNames.REFERER_PARAM + "=" + URLS[random.nextInt(URLS.length)] + "]");
-                argsList.add("[" + ParamNames.USER_ID_PARAM + "=" + user + "]");
+                argsList.add("[resource_id=" + RESOURCES[random.nextInt(RESOURCES.length)] + "]");
+                argsList.add("[sesion_id=" + sessionId + "]");
+                argsList.add("[ip_address=" + IPADDRESSES[random.nextInt(IPADDRESSES.length)] + "]");
+                argsList.add("[url=" + URLS[random.nextInt(URLS.length)] + "]");
+                argsList.add("[http_referer=" + URLS[random.nextInt(URLS.length)] + "]");
+                argsList.add("[user_id=" + user + "]");
 
                 String[] args = new String[argsList.size()];
                 args = argsList.toArray(args);
